+++ src/main/java/commoble/froglins/ai/JumpSometimesGoal.java
@@ -5,8 +5,6 @@ import java.util.EnumSet;
 import commoble.froglins.FroglinEntity;
 import net.minecraft.world.entity.ai.goal.Goal;
 
-import net.minecraft.world.entity.ai.goal.Goal.Flag;
-
 public class JumpSometimesGoal extends Goal
 {
 	private final FroglinEntity froglin;